Moving magnolia stuff to scala-2 src
